+++ libsolidity/parsing/Parser.cpp
@@ -821,7 +821,6 @@ ASTPointer<VariableDeclaration> Parser::parseVariableDeclaration(
 		value,
 		visibility,
 		documentation,
-		_options.isStateVariable,
 		isIndexed,
 		mutability,
 		overrides,