+++ solc/CommandLineInterface.cpp
@@ -172,6 +172,8 @@ static string const g_strOutputDir = "output-dir";
 static string const g_strOverwrite = "overwrite";
 static string const g_strRevertStrings = "revert-strings";
 static string const g_strStorageLayout = "storage-layout";
+static string const g_strStopAfter = "stop-after";
+static string const g_strParsing = "parsing";
 
 /// Possible arguments to for --revert-strings
 static set<string> const g_revertStringsArgs
@@ -344,6 +346,17 @@ static bool needsHumanTargetedStdout(po::variables_map const& _args)
 		if (_args.count(arg))
 			return true;
 	return false;
+}
+
+bool checkMutuallyExclusive(boost::program_options::variables_map const& args, std::string const& _optionA, std::string const& _optionB)
+{
+	if (args.count(_optionA) && args.count(_optionB))
+	{
+		serr() << "Option " << _optionA << " and " << _optionB << " are mutually exclusive." << endl;
+		return false;
+	}
+
+	return true;
 }
 
 void CommandLineInterface::handleBinary(string const& _contract)
@@ -830,6 +843,11 @@ General Information)").c_str(),
 			po::value<string>()->value_name(boost::join(g_revertStringsArgs, ",")),
 			"Strip revert (and require) reason strings or add additional debugging information."
 		)
+		(
+			g_strStopAfter.c_str(),
+			po::value<string>()->value_name("stage"),
+			"Stop execution after the given compiler stage. Valid options: \"parsing\"."
+		)
 	;
 	desc.add(outputOptions);
 
@@ -1030,11 +1048,23 @@ General Information)").c_str(),
 		return false;
 	}
 
-	if (m_args.count(g_argColor) && m_args.count(g_argNoColor))
-	{
-		serr() << "Option " << g_argColor << " and " << g_argNoColor << " are mutualy exclusive." << endl;
+	if (!checkMutuallyExclusive(m_args, g_argColor, g_argNoColor))
 		return false;
-	}
+
+	static vector<string> const conflictingWithStopAfter{
+		g_argBinary,
+		g_argIR,
+		g_argIROptimized,
+		g_argEwasm,
+		g_argGas,
+		g_argAsm,
+		g_argAsmJson,
+		g_argOpcodes
+	};
+
+	for (auto& option: conflictingWithStopAfter)
+		if (!checkMutuallyExclusive(m_args, g_strStopAfter, option))
+			return false;
 
 	m_coloredOutput = !m_args.count(g_argNoColor) && (isatty(STDERR_FILENO) || m_args.count(g_argColor));
 
@@ -1170,6 +1200,17 @@ bool CommandLineInterface::processInput()
 				filesystem_path.remove_filename();
 			m_allowedDirectories.push_back(filesystem_path);
 		}
+	}
+
+	if (m_args.count(g_strStopAfter))
+	{
+		if (m_args[g_strStopAfter].as<string>() != "parsing")
+		{
+			serr() << "Valid options for --" << g_strStopAfter << " are: \"parsing\".\n";
+			return false;
+		}
+		else
+			m_stopAfter = CompilerStack::State::Parsed;
 	}
 
 	vector<string> const exclusiveModes = {
@@ -1451,7 +1492,7 @@ bool CommandLineInterface::processInput()
 				m_compiler->setParserErrorRecovery(true);
 		}
 
-		bool successful = m_compiler->compile();
+		bool successful = m_compiler->compile(m_stopAfter);
 
 		for (auto const& error: m_compiler->errors())
 		{
@@ -1599,7 +1640,7 @@ void CommandLineInterface::handleCombinedJSON()
 		output[g_strSources] = Json::Value(Json::objectValue);
 		for (auto const& sourceCode: m_sourceCodes)
 		{
-			ASTJsonConverter converter(legacyFormat, m_compiler->sourceIndices());
+			ASTJsonConverter converter(legacyFormat, m_compiler->state(), m_compiler->sourceIndices());
 			output[g_strSources][sourceCode.first] = Json::Value(Json::objectValue);
 			output[g_strSources][sourceCode.first]["AST"] = converter.toJson(m_compiler->ast(sourceCode.first));
 		}
@@ -1631,18 +1672,6 @@ void CommandLineInterface::handleAst(string const& _argStr)
 		vector<ASTNode const*> asts;
 		for (auto const& sourceCode: m_sourceCodes)
 			asts.push_back(&m_compiler->ast(sourceCode.first));
-		map<ASTNode const*, evmasm::GasMeter::GasConsumption> gasCosts;
-		for (auto const& contract: m_compiler->contractNames())
-			if (m_compiler->compilationSuccessful())
-				if (auto const* assemblyItems = m_compiler->runtimeAssemblyItems(contract))
-				{
-					auto ret = GasEstimator::breakToStatementLevel(
-						GasEstimator(m_evmVersion).structuralEstimation(*assemblyItems, asts),
-						asts
-					);
-					for (auto const& it: ret)
-						gasCosts[it.first] += it.second;
-				}
 
 		bool legacyFormat = !m_args.count(g_argAstCompactJson);
 		if (m_args.count(g_argOutputDir))
@@ -1651,7 +1680,7 @@ void CommandLineInterface::handleAst(string const& _argStr)
 			{
 				stringstream data;
 				string postfix = "";
-				ASTJsonConverter(legacyFormat, m_compiler->sourceIndices()).print(data, m_compiler->ast(sourceCode.first));
+				ASTJsonConverter(legacyFormat, m_compiler->state(), m_compiler->sourceIndices()).print(data, m_compiler->ast(sourceCode.first));
 				postfix += "_json";
 				boost::filesystem::path path(sourceCode.first);
 				createFile(path.filename().string() + postfix + ".ast", data.str());
@@ -1663,7 +1692,7 @@ void CommandLineInterface::handleAst(string const& _argStr)
 			for (auto const& sourceCode: m_sourceCodes)
 			{
 				sout() << endl << "======= " << sourceCode.first << " =======" << endl;
-				ASTJsonConverter(legacyFormat, m_compiler->sourceIndices()).print(sout(), m_compiler->ast(sourceCode.first));
+				ASTJsonConverter(legacyFormat, m_compiler->state(), m_compiler->sourceIndices()).print(sout(), m_compiler->ast(sourceCode.first));
 			}
 		}
 	}
@@ -2014,7 +2043,10 @@ void CommandLineInterface::outputCompilationResults()
 	if (m_args.count(g_argAstBoogie))
 		handleBoogie();
 
-	if (!m_compiler->compilationSuccessful())
+	if (
+		!m_compiler->compilationSuccessful() &&
+		m_stopAfter == CompilerStack::State::CompilationSuccessful
+	)
 	{
 		serr() << endl << "Compilation halted after AST generation due to errors." << endl;
 		return;